+++ jib-cli/src/test/java/com/google/cloud/tools/jib/api/HttpRequestTester.java
@@ -14,7 +14,7 @@
  * the License.
  */
 
-package com.google.cloud.tools.jib.maven;
+package com.google.cloud.tools.jib.api;
 
 import com.google.cloud.tools.jib.blob.Blobs;
 import java.io.IOException;
@@ -24,15 +24,28 @@ import java.net.URL;
 import javax.annotation.Nullable;
 import org.junit.Assert;
 
-/** Verifies the response of HTTP GET. */
-class HttpGetVerifier {
+/** Test helpers for making HTTP requests. */
+public class HttpRequestTester {
 
   /**
    * Verifies the response body. Repeatedly tries {@code url} at the interval of .5 seconds for up
    * to 20 seconds until getting OK HTTP response code.
    */
-  static void verifyBody(String expectedBody, URL url) throws InterruptedException {
+  public static void verifyBody(String expectedBody, URL url) throws InterruptedException {
     Assert.assertEquals(expectedBody, getContent(url));
+  }
+
+  /** Fetches the host to use for the http request. */
+  public static String fetchDockerHostForHttpRequest() {
+    if (System.getenv("KOKORO_JOB_CLUSTER") != null
+        && System.getenv("KOKORO_JOB_CLUSTER").equals("MACOS_EXTERNAL")) {
+      return System.getenv("DOCKER_IP");
+    } else if (System.getenv("KOKORO_JOB_CLUSTER") != null
+        && System.getenv("KOKORO_JOB_CLUSTER").equals("GCP_UBUNTU_DOCKER")) {
+      return System.getenv("DOCKER_IP_UBUNTU");
+    } else {
+      return "localhost";
+    }
   }
 
   @Nullable
@@ -52,4 +65,6 @@ class HttpGetVerifier {
     }
     return null;
   }
+
+  private HttpRequestTester() {}
 }