+++ jib-core/src/main/java/com/google/cloud/tools/jib/configuration/BuildContext.java
@@ -548,22 +548,23 @@ public class BuildContext implements Closeable {
   }
 
   /**
-   * Creates a new {@link RegistryClient.Factory} for the base image with fields from the build
-   * configuration. The server URL is derived from the base {@link ImageConfiguration}.
+   * Creates a new {@link com.google.cloud.tools.jib.registry.RegistryClient.Factory} for the base
+   * image with fields from the build configuration. The server URL is derived from the base {@link
+   * ImageConfiguration}.
    *
-   * @return a new {@link RegistryClient.Factory}
+   * @return a new {@link com.google.cloud.tools.jib.registry.RegistryClient.Factory}
    */
   public RegistryClient.Factory newBaseImageRegistryClientFactory() {
     return newBaseImageRegistryClientFactory(baseImageConfiguration.getImageRegistry());
   }
 
   /**
-   * Creates a new {@link RegistryClient.Factory} for the base image repository on the registry
-   * {@code serverUrl}. Compared to @link #newBaseImageRegistryClientFactory()), this method is
-   * useful to try a mirror.
+   * Creates a new {@link com.google.cloud.tools.jib.registry.RegistryClient.Factory} for the base
+   * image repository on the registry {@code serverUrl}. Compared to @link
+   * #newBaseImageRegistryClientFactory()), this method is useful to try a mirror.
    *
    * @param serverUrl the server URL for the registry (for example, {@code gcr.io})
-   * @return a new {@link RegistryClient.Factory}
+   * @return a new {@link com.google.cloud.tools.jib.registry.RegistryClient.Factory}
    */
   public RegistryClient.Factory newBaseImageRegistryClientFactory(String serverUrl) {
     return RegistryClient.factory(
@@ -572,10 +573,10 @@ public class BuildContext implements Closeable {
   }
 
   /**
-   * Creates a new {@link RegistryClient.Factory} for the target image with fields from the build
-   * configuration.
+   * Creates a new {@link com.google.cloud.tools.jib.registry.RegistryClient.Factory} for the target
+   * image with fields from the build configuration.
    *
-   * @return a new {@link RegistryClient.Factory}
+   * @return a new {@link com.google.cloud.tools.jib.registry.RegistryClient.Factory}
    */
   public RegistryClient.Factory newTargetImageRegistryClientFactory() {
     // if base and target are on the same registry, try enabling cross-repository mounts