+++ jib-plugins-common/src/test/java/com/google/cloud/tools/jib/plugins/common/UpdateCheckerTest.java
@@ -68,8 +68,9 @@ public class UpdateCheckerTest {
   }
 
   @Test
-  public void testPerformUpdateCheck_newVersionFound() throws IOException {
+  public void testPerformUpdateCheck_newVersionFound() throws IOException, InterruptedException {
     Instant before = Instant.now();
+    Thread.sleep(100);
     setupLastUpdateCheck();
     Optional<String> message =
         UpdateChecker.performUpdateCheck(
@@ -99,8 +100,9 @@ public class UpdateCheckerTest {
   }
 
   @Test
-  public void testPerformUpdateCheck_onLatest() throws IOException {
+  public void testPerformUpdateCheck_onLatest() throws IOException, InterruptedException {
     Instant before = Instant.now();
+    Thread.sleep(100);
     setupLastUpdateCheck();
     Optional<String> message =
         UpdateChecker.performUpdateCheck(
@@ -115,8 +117,9 @@ public class UpdateCheckerTest {
   }
 
   @Test
-  public void testPerformUpdateCheck_noLastUpdateCheck() throws IOException {
+  public void testPerformUpdateCheck_noLastUpdateCheck() throws IOException, InterruptedException {
     Instant before = Instant.now();
+    Thread.sleep(100);
     Optional<String> message =
         UpdateChecker.performUpdateCheck(
             configDir, "1.0.2", testWebServer.getEndpoint(), "tool-name", ignored -> {});
@@ -129,9 +132,11 @@ public class UpdateCheckerTest {
   }
 
   @Test
-  public void testPerformUpdateCheck_emptyLastUpdateCheck() throws IOException {
+  public void testPerformUpdateCheck_emptyLastUpdateCheck()
+      throws IOException, InterruptedException {
     Files.createFile(configDir.resolve("lastUpdateCheck"));
     Instant before = Instant.now();
+    Thread.sleep(100);
     Optional<String> message =
         UpdateChecker.performUpdateCheck(
             configDir, "1.0.2", testWebServer.getEndpoint(), "tool-name", ignored -> {});
@@ -164,8 +169,9 @@ public class UpdateCheckerTest {
   }
 
   @Test
-  public void testPerformUpdateCheck_badLastUpdateTime() throws IOException {
+  public void testPerformUpdateCheck_badLastUpdateTime() throws IOException, InterruptedException {
     Instant before = Instant.now();
+    Thread.sleep(100);
     Files.write(
         configDir.resolve("lastUpdateCheck"), "bad timestamp".getBytes(StandardCharsets.UTF_8));
     Optional<String> message =